+++ packages/alfa-css/src/value/image/index.ts
@@ -1,5 +1,3 @@
 export * from "./gradient";
-export * from "./gradient-linear";
-export * from "./gradient-radial";
 export * from "./image";
 export * from "./url";
