+++ packages/alfa-css/tsconfig.json
@@ -44,9 +44,13 @@
     "src/value/color/named.ts",
     "src/value/color/rgb.ts",
     "src/value/color/system.ts",
-    "src/value/image/gradient.ts",
-    "src/value/image/gradient-linear.ts",
-    "src/value/image/gradient-radial.ts",
+    "src/value/image/gradient/gradient.ts",
+    "src/value/image/gradient/hint.ts",
+    "src/value/image/gradient/index.ts",
+    "src/value/image/gradient/item.ts",
+    "src/value/image/gradient/linear.ts",
+    "src/value/image/gradient/radial.ts",
+    "src/value/image/gradient/stop.ts",
     "src/value/image/image.ts",
     "src/value/image/index.ts",
     "src/value/image/url.ts",
